requirements.txt
streamlit>=1.33.0
pandas>=2.0.0
plotly>=5.18.0
pyarrow>=14.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
supabase>=2.4.0
psycopg2-binary>=2.9.9
